pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-cli"
version = "0.1.0"
description = "Context walker for the Forge L0-L5 spec system"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
forge = "cli.forge:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
